macos: drive utun as a point-to-point (non-ethernet) link
The old tap_darwin.c faked an Ethernet link on top of utun: it
synthesized Ethernet headers around IP packets and answered the
stack's ARP requests in-process (request intercepted in send, reply
bounced back through a pending-frame slot on the next poll). That ARP
simulation made the first TCP exchange on macOS timing-sensitive, and
the FIFO space accounting fix (F-8521) exposed it as a hard hang of
test-wolfssl (macos_test 3-minute timeout; identical code passes once
the timing is perturbed).

utun is a native L3 point-to-point device: it carries IP packets
prefixed with a 4-byte address family field and needs no ARP at all.
Register the link as non_ethernet so wolfIP supplies the dummy link
header on receive and strips it on transmit (wolfIP_ll_send_frame),
and skips ARP for this interface - the same driver contract as the
Linux TUN driver.

- utun_poll: read datagram, verify/strip the AF_INET prefix, hand the
  raw IP packet to the stack
- utun_send: prepend the AF_INET prefix, single write so prefix and
  packet arrive as one datagram
- drop the ARP interception, pending-frame buffer, Ethernet
  synthesis, peer MAC tables and the driver-internal locks (poll and
  send run on the stack's single thread)
- rename tap_darwin.c to utun_darwin.c; Makefile selects it on Darwin

Function entry point stays tap_init() so the test harnesses need no
changes.
